Added "Upgradeable" Column
I was using this plugin to help generate a report I was writing in
excel for my boss then I realized that I should just and the check to
the program so I can just hit update and print out the results when
needed.   Don't know if you'll add it to the core but I like.  Just an
idea....